move common args to separate module
